Speed up creating the table layout
tabulator is very slow when the table is rendered and columns are added, removed, or moved.
These changes lead to everything getting re-rendered all the time.

Instead of applying these changes one by one simply replace all column specs when changing
something related to the columns (order, visibility).

While at it, this also fixes the column header translations not changing when the language changes.
We can simply re-build the columns in that case.